cleaned up tests. Lag on LSA conversion tests passing
