+++ featuretools/tests/feature_discovery/test_feature_discovery.py
@@ -6,14 +6,8 @@ from woodwork.column_schema import ColumnSchema
 from woodwork.logical_types import Boolean, Datetime, Double, Ordinal
 
 from featuretools.entityset.entityset import EntitySet
-from featuretools.feature_base import Feature as OldFeature
-from featuretools.feature_base import (
-    IdentityFeature,
-    TransformFeature,
-)
 from featuretools.feature_discovery.feature_discovery import (
     feature_to_keys,
-    features_from_primitive,
     generate_hashing_keys_from_column_schema,
     get_features,
     get_matching_features,
@@ -24,11 +18,8 @@ from featuretools.feature_discovery.feature_discovery import (
 )
 from featuretools.feature_discovery.type_defs import (
     Feature,
-    convert_feature_list_to_featurebase_list,
-    convert_feature_to_featurebase2,
 )
 from featuretools.primitives import (
-    LSA,
     Absolute,
     AddNumeric,
     DateFirstEvent,
@@ -416,193 +407,3 @@ def test_compare_dfs(col_defs, primitives):
     )
 
     assert feature_names_old == feature_names_new
-
-
-def test_lag_on_lsa():
-    col_defs = [
-        ("idx", "Double", {"index"}),
-        ("t_idx", "Datetime", {"time_index"}),
-        ("f_1", "NaturalLanguage"),
-    ]
-    df = generate_fake_dataframe(
-        col_defs=col_defs,
-    )
-
-    primitives = [LSA]
-
-    es = EntitySet(id="test")
-    es.add_dataframe(df, df.ww.name, index="idx")
-
-    features_old = dfs(
-        entityset=es,
-        target_dataframe_name=df.ww.name,
-        trans_primitives=primitives,
-        features_only=True,
-    )
-    assert len(features_old) == 1
-
-    feature = features_old[0]
-    time_index_feature = OldFeature(df.ww["t_idx"])
-
-    base_features = [feature, time_index_feature]
-
-    lag_instance = Lag(periods=2)
-
-    with pytest.raises(ValueError, match="Cannot stack on whole multi-output feature."):
-        OldFeature(base_features, primitive=lag_instance)
-
-    split_lsa = [feature[i] for i in range(feature.number_output_features)]
-    lagged_lsa_features = []
-    for f in split_lsa:
-        base_features = [f, time_index_feature]
-        lagged_lsa_features.append(OldFeature(base_features, primitive=lag_instance))
-
-    origin_features = schema_to_features(df.ww.schema)
-
-    feature_groups = group_features(origin_features)
-    lsa_features = features_from_primitive(feature_groups, LSA())
-
-    # Confirm 2 LSA features are properly converted into a single FeatureBase
-    # LSA feature with proper names
-    converted_features = convert_feature_list_to_featurebase_list(df, lsa_features)
-
-    assert len(converted_features) == 1
-    lsa_feature = converted_features[0]
-    assert lsa_feature.get_feature_names() == ["LSA(f_1)[0]", "LSA(f_1)[1]"]
-
-    time_index_feature = [x for x in origin_features if x.name == "t_idx"][0]
-    feature_groups = group_features(lsa_features + [time_index_feature])
-    lag_features = features_from_primitive(feature_groups, lag_instance)
-    assert len(lag_features) == 2
-    assert set([x.get_name() for x in lag_features]) == set(
-        ["LAG(LSA(f_1)[0], t_idx, periods=2)", "LAG(LSA(f_1)[1], t_idx, periods=2)"],
-    )
-
-    converted_features = convert_feature_list_to_featurebase_list(df, lag_features)
-
-    assert set([x.get_name() for x in converted_features]) == set(
-        ["LAG(LSA(f_1)[0], t_idx, periods=2)", "LAG(LSA(f_1)[1], t_idx, periods=2)"],
-    )
-
-
-def test_origin_feature_conversion():
-    df = generate_fake_dataframe(
-        col_defs=[("idx", "Double", {"index"}), ("f_1", "Double")],
-    )
-    # TODO(dreed): don't like how I have to make an entityset
-    es = EntitySet(id="test")
-    es.add_dataframe(df, df.ww.name)
-
-    origin_features = schema_to_features(df.ww.schema)
-    f_1 = [f for f in origin_features if f.name == "f_1"][0]
-    fb = convert_feature_to_featurebase2(f_1, df)
-
-    assert isinstance(fb, IdentityFeature)
-    assert fb.get_name() == "f_1"
-
-    f_1.rename("new name")
-
-    df.ww.rename({"f_1": "new name"}, inplace=True)
-    fb = convert_feature_to_featurebase2(f_1, df)
-
-    assert isinstance(fb, IdentityFeature)
-    assert fb.get_name() == "new name"
-
-
-def test_stacked_feature_conversion():
-    df = generate_fake_dataframe(
-        col_defs=[("idx", "Double", {"index"}), ("f_1", "Double")],
-    )
-    # TODO(dreed): don't like how I have to make an entityset
-    es = EntitySet(id="test")
-    es.add_dataframe(df, df.ww.name)
-
-    origin_features = schema_to_features(df.ww.schema)
-    f_1 = [f for f in origin_features if f.name == "f_1"][0]
-    fc = my_dfs([f_1], [Absolute()])
-
-    f_2 = [f for f in fc.all_features if f.name == "ABSOLUTE(f_1)"][0]
-
-    fb = convert_feature_to_featurebase2(f_2, df)
-
-    assert isinstance(fb, TransformFeature)
-    assert fb.get_name() == "ABSOLUTE(f_1)"
-    assert len(fb.base_features) == 1
-    assert fb.base_features[0].get_name() == "f_1"
-
-    f_2.rename("f_2")
-    fb = convert_feature_to_featurebase2(f_2, df)
-
-    assert isinstance(fb, TransformFeature)
-    assert fb.get_name() == "f_2"
-    assert len(fb.base_features) == 1
-    assert fb.base_features[0].get_name() == "f_1"
-
-
-def test_multi_output_to_featurebase():
-    df = generate_fake_dataframe(
-        col_defs=[
-            ("idx", "Double", {"index"}),
-            ("f_1", "NaturalLanguage"),
-        ],
-    )
-
-    # TODO(dreed): don't like how I have to make an entityset
-    es = EntitySet(id="test")
-    es.add_dataframe(df, df.ww.name)
-
-    # primitive_instance = LSA()
-    origin_features = schema_to_features(df.ww.schema)
-    f_1 = [f for f in origin_features if f.name == "f_1"][0]
-    fc = my_dfs([f_1], [LSA()])
-
-    lsa_features = [f for f in fc.all_features if f.get_primitive_name() == "lsa"]
-    assert len(lsa_features) == 2
-
-    fb = convert_feature_to_featurebase2(lsa_features[0], df)
-    assert isinstance(fb, TransformFeature)
-    assert fb.get_name() == "LSA(f_1)"
-    assert len(fb.base_features) == 1
-    assert set(fb.get_feature_names()) == set(["LSA(f_1)[0]", "LSA(f_1)[1]"])
-    assert fb.base_features[0].get_name() == "f_1"
-
-    lsa_features[0].rename("f_2")
-    lsa_features[1].rename("f_3")
-
-    fb = convert_feature_to_featurebase2(lsa_features[0], df)
-    assert isinstance(fb, TransformFeature)
-    assert len(fb.base_features) == 1
-    assert set(fb.get_feature_names()) == set(["f_2", "f_3"])
-    assert fb.base_features[0].get_name() == "f_1"
-
-
-def test_stacking_on_multioutput_to_featurebase():
-    col_defs = [
-        ("idx", "Double", {"index"}),
-        ("t_idx", "Datetime", {"time_index"}),
-        ("f_1", "NaturalLanguage"),
-    ]
-    df = generate_fake_dataframe(
-        col_defs=col_defs,
-    )
-    # TODO(dreed): don't like how I have to make an entityset
-    es = EntitySet(id="test")
-    es.add_dataframe(df, df.ww.name)
-
-    origin_features = schema_to_features(df.ww.schema)
-    time_index_feature = [f for f in origin_features if f.name == "t_idx"][0]
-    f_1 = [f for f in origin_features if f.name == "f_1"][0]
-
-    fc = my_dfs([f_1], [LSA()])
-    lsa_features = [f for f in fc.all_features if f.get_primitive_name() == "lsa"]
-    assert len(lsa_features) == 2
-
-    fc = my_dfs(lsa_features + [time_index_feature], [Lag(periods=2)])
-    lag_features = [f for f in fc.all_features if f.get_primitive_name() == "lag"]
-    assert len(lag_features) == 2
-
-    fb = convert_feature_to_featurebase2(lag_features[0], df)
-
-    assert isinstance(fb, TransformFeature)
-    assert fb.get_name() == lag_features[0].get_name()
-    assert len(fb.base_features) == 2
